Speed up CLI access.
It was taking too long to access the CLI, even to call the usage statement.
Changed imports in __init__.py to only run when the module is called from
inside the Python interpreter, not from CLI.
Added more specific arguments to individual modules.

Split command_line.py into separate functions to handle arguments and run
the program so that packages are not loaded unless they are needed.

Updated to v0.4.0

 Changes to be committed:
	modified:   ibdpainting/__init__.py
	modified:   ibdpainting/command_line.py
	modified:   ibdpainting/ibd_table.py
	modified:   ibdpainting/load_genotype_data.py