Load HDF5 files one window at a time.
A substantial reimplementation of how to load the data.
ibd_table now imports genotype data window by window, which is much faster.
I moved the functions for pairwise_distance and heterozygosity from the
geneticDistance class to the ibd_table module.
I removed load_genotype_data and geneticDistance, as they are now redundant.

This also fixes a very pernicious bug in calculating heterozygosity which
was passing on masked arrays, which plot_ibd_table was treating as read-only.

I also changed plot_ibd_table to swap the axes of the output plots. so that
genetic distances are now shown as similarities. I think this is more
intuitive now that I also plot heterozygous calls, because it means that
'higher means a better match'.

 On branch main
 Changes to be committed:
	deleted:    ibdpainting/geneticDistance.py
	modified:   ibdpainting/ibd_scores.py
	modified:   ibdpainting/ibd_table.py
	deleted:    ibdpainting/load_genotype_data.py
	modified:   ibdpainting/plot_ibd_table.py
	deleted:    tests/test_genotypeDistance.py
	modified:   tests/test_ibd_table.py
	deleted:    tests/test_load_genotypes.py